Add "groupid" param to "create report" & "create map" links
When viewing a group’s library, add the group_id to requests to create
a new map or report so that the mapping environment knows where they’ve
come from. Also allow other plugins to filter the text of the “Create a
New Map” and “Create a New Report” links.